+++ R/Ano.R
@@ -21,10 +21,10 @@
 #'clim <- Clim(sampleData$mod, sampleData$obs)
 #'ano_exp <- Ano(sampleData$mod, clim$clim_exp)
 #'ano_obs <- Ano(sampleData$obs, clim$clim_obs)
-#'\donttest{
+#'\dontrun{
 #'PlotAno(ano_exp, ano_obs, startDates, 
 #'        toptitle = 'Anomaly', ytitle = c('K', 'K', 'K'), 
-#'        legends = 'ERSST', biglab = FALSE, fileout = 'tos_ano.png')
+#'        legends = 'ERSST', biglab = FALSE)
 #'}
 #'@import multiApply
 #'@export