+++ R/Ano_CrossValid.R
@@ -47,7 +47,7 @@
 #'\dontrun{
 #'PlotAno(anomalies$exp, anomalies$obs, startDates, 
 #'        toptitle = paste('anomalies'), ytitle = c('K', 'K', 'K'), 
-#'        legends = 'ERSST', biglab = FALSE, fileout = 'tos_ano_crossvalid.eps')
+#'        legends = 'ERSST', biglab = FALSE)
 #'}
 #'@import multiApply
 #'@importFrom ClimProjDiags Subset