+++ R/Spread.R
@@ -54,27 +54,27 @@
 #'                                                   name = 'member')
 #'spread <- Spread(smooth_ano_exp_m_sub, compute_dim = c('member', 'sdate'))
 #'
-#'\donttest{
+#'\dontrun{
 #'PlotVsLTime(Reorder(spread$iqr, c('dataset', 'stats', 'ftime')), 
 #'            toptitle = "Inter-Quartile Range between ensemble members",
 #'            ytitle = "K", monini = 11, limits = NULL, 
 #'            listexp = c('CMIP5 IC3'), listobs = c('ERSST'), biglab = FALSE, 
-#'            hlines = c(0), fileout = 'tos_iqr.png')
+#'            hlines = c(0))
 #'PlotVsLTime(Reorder(spread$maxmin, c('dataset', 'stats', 'ftime')), 
 #'            toptitle = "Maximum minus minimum of the members", 
 #'            ytitle = "K", monini = 11, limits = NULL, 
 #'            listexp = c('CMIP5 IC3'), listobs = c('ERSST'), biglab = FALSE, 
-#'            hlines = c(0), fileout = 'tos_maxmin.png')
+#'            hlines = c(0))
 #'PlotVsLTime(Reorder(spread$sd, c('dataset', 'stats', 'ftime')), 
 #'            toptitle = "Standard deviation of the members", 
 #'            ytitle = "K", monini = 11, limits = NULL, 
 #'            listexp = c('CMIP5 IC3'), listobs = c('ERSST'), biglab = FALSE, 
-#'            hlines = c(0), fileout = 'tos_sd.png')
+#'            hlines = c(0))
 #'PlotVsLTime(Reorder(spread$mad, c('dataset', 'stats', 'ftime')), 
 #'            toptitle = "Median Absolute Deviation of the members",
 #'            ytitle = "K", monini = 11, limits = NULL, 
 #'            listexp = c('CMIP5 IC3'), listobs = c('ERSST'), biglab = FALSE, 
-#'            hlines = c(0), fileout = 'tos_mad.png')
+#'            hlines = c(0))
 #'}
 #'
 #'@import multiApply
@@ -114,12 +114,12 @@ Spread <- function(data, compute_dim = 'member', na.rm = TRUE,
     stop("Parameter 'conf' must be one logical value.")
   }
   ## conf.lev
-  if (!is.numeric(conf.lev) | conf.lev < 0 | conf.lev > 1 | length(conf.lev) > 1) {
+  if (!is.numeric(conf.lev) | any(conf.lev < 0) | any(conf.lev > 1) | length(conf.lev) > 1) {
     stop("Parameter 'conf.lev' must be a numeric number between 0 and 1.")
   }
   ## ncores
   if (!is.null(ncores)) {
-    if (!is.numeric(ncores) | ncores %% 1 != 0 | ncores <= 0 |
+    if (!is.numeric(ncores) | any(ncores %% 1 != 0) | any(ncores <= 0) |
       length(ncores) > 1) {
       stop("Parameter 'ncores' must be a positive integer.")
     }